Add resulting expectation suite, example stub data file
